+++ tests/test_cli_update_recipe.py
@@ -17,7 +17,7 @@ def test_cli_update_recipe_no_recipe():
 
 @pytest.mark.usefixtures("token_environ")
 @mock.patch("requests.patch")
-def test_cli_update_recipe(mock_request_patch, MockResponse):
+def test_cli_update_recipe_201(mock_request_patch, MockResponse):
     runner = CliRunner()
 
     # sends expected request
@@ -32,10 +32,25 @@ def test_cli_update_recipe(mock_request_patch, MockResponse):
 
 @pytest.mark.usefixtures("token_environ")
 @mock.patch("requests.patch")
+def test_cli_update_recipe_204(mock_request_patch, MockResponse):
+    runner = CliRunner()
+
+    # sends expected request
+    mock_request_patch.return_value = MockResponse("", status_code=204)
+    result = runner.invoke(update_recipe, ["test.id", "tests/fixtures/recipe.json"])
+    mock_request_patch.assert_called_with(
+        "https://api.mapbox.com/tilesets/v1/test.id/recipe?access_token=fake-token",
+        json={"minzoom": 0, "maxzoom": 10, "layer_name": "test_layer"},
+    )
+    assert result.exit_code == 0
+
+
+@pytest.mark.usefixtures("token_environ")
+@mock.patch("requests.patch")
 def test_cli_update_recipe2(mock_request_patch, MockResponse):
     runner = CliRunner()
 
-    mock_request_patch.return_value = MockResponse({}, status_code=201)
+    mock_request_patch.return_value = MockResponse("", status_code=204)
     # Provides the flag --token
     result = runner.invoke(
         update_recipe,